Move truss config to base /truss path and move main.go to /cmd/truss
