+++ Makefile
@@ -3,6 +3,9 @@
 # Build native Truss by default.
 default: truss
 
+dependencies: 
+	go install github.com/golang/protobuf/protoc-gen-go
+
 # Generate go files containing the all template files in []byte form
 gobindata:
 	go generate github.com/TuneLab/go-truss/gengokit/template
@@ -10,8 +13,8 @@ gobindata:
 
 # Install truss and protoc-gen-truss-protocast
 truss: gobindata
-	go install github.com/TuneLab/go-truss/protoc-gen-truss-protocast
-	go install github.com/TuneLab/go-truss/truss
+	go install github.com/TuneLab/go-truss/cmd/protoc-gen-truss-protocast
+	go install github.com/TuneLab/go-truss/cmd/truss
 
 # Run the go tests and the truss integration tests
 test: test-go test-integration
@@ -20,13 +23,8 @@ test-go:
 	go test -v ./...
 
 test-integration:
-	$(MAKE) -C truss test-integration
-
-# Run the go non-vendored unit tests
-test-nv:
-	go test -v ./deftree/... ./gendoc/... ./gengokit/... \
-		./protoc-gen-truss-protocast/... ./truss/...
+	$(MAKE) -C cmd/_integration-tests
 
 # Removes generated code from tests
-testclean: 
-	$(MAKE) -C truss testclean
+testclean:
+	$(MAKE) -C _integration-tests clean
