+++ cmd/_integration-tests/cli/test-service-definitions/1-basic/basic.proto
