+++ cmd/_integration-tests/cli/test-service-definitions/1-multifile/basic.proto
