+++ cmd/_integration-tests/cli/test-service-definitions/2-repeated/repeated.proto
