+++ cmd/_integration-tests/cli/test-service-definitions/3-nested/nested.proto
