+++ cmd/_integration-tests/cli/test-service-definitions/4-maps/maps.proto
