+++ cmd/_integration-tests/transport/handlers/server/server_handler.go
@@ -6,7 +6,7 @@ package handler
 import (
 	"golang.org/x/net/context"
 
-	pb "github.com/TuneLab/go-truss/truss/_integration-tests/transport/transport-service"
+	pb "github.com/TuneLab/go-truss/cmd/_integration-tests/transport/transport-service"
 )
 
 // NewService returns a naïve, stateless implementation of Service.
