+++ cmd/truss/README.md
@@ -1,6 +1,5 @@
 # truss
 
-
 The `truss` binary reads in gRPC files that define a single gRPC *service* and outputs:
 
 1. Markdown and html documentation based on comments in the gRPC files.
@@ -15,18 +14,4 @@ The `truss` binary reads in gRPC files that define a single gRPC *service* and o
 	- http/json transport (including all encoding/decoding)
 	- no-op handler methods for each *service* rpc, ready for marshalling command line arguments into a request object
 4. An web based api explorer (through naive swagger generation)
-
-## Install
-
-```
-$ go get -u -v github.com/TuneLab/go-truss/...
-
-```
-
-## Requirements
-
-`truss` must:
-- Be invoked from some directory within your `$GOPATH/src`
-- Be passed `.proto` file paths that:
-	- Are withing the current directory
