--- truss/Makefile
@@ -1,9 +1,0 @@
-default:
-	$(MAKE) -C ..
-
-test-integration: testclean
-	$(MAKE) -C _integration-tests
-
-# Removes generated code from tests
-testclean:
-	$(MAKE) -C _integration-tests clean
