+++ truss/truss.go
